Refactor helpers to conditionally use pgzip
