+++ agent/command/archive_zip_create.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/evergreen-ci/evergreen/agent/internal"
 	"github.com/evergreen-ci/evergreen/agent/internal/client"
-	agentutil "github.com/evergreen-ci/evergreen/agent/util"
 	"github.com/evergreen-ci/evergreen/util"
 	"github.com/mholt/archiver/v3"
 	"github.com/mitchellh/mapstructure"
@@ -72,7 +71,7 @@ func (c *zipArchiveCreate) Execute(ctx context.Context,
 		c.Target = getWorkingDirectory(conf, c.Target)
 	}
 
-	files, err := agentutil.FindContentsToArchive(ctx, c.SourceDir, c.Include, c.ExcludeFiles)
+	files, _, err := FindContentsToArchive(ctx, c.SourceDir, c.Include, c.ExcludeFiles)
 	if err != nil {
 		return errors.Wrap(err, "finding files to archive")
 	}
